+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/Changelog.rst
@@ -15,6 +15,10 @@ Unreleased
 Fixed
 *****
 - When kicking players, we need to use DisconnectFromServer (which as expected disconnects from the server ) and not Disconnect, which just disconnects from scene. (Should probably be renamed DisconnectFromScene...)
+Added
+*****
+- Kick now supports the filter */authenticated which enables kicking only authenticated users. (queued users are not kicked) 
+- Kick now supports the filter */!authenticated which enables kicking only non authenticated users. (queued users are kicked but not authenticated users for instance.) 
 
 5.0.1.5
 -------